refactor: migrate to common.storage package
