+++ AnkiDroid/src/test/java/com/ichi2/anki/StorageDecisionGateTest.kt
@@ -5,9 +5,10 @@ package com.ichi2.anki
 import androidx.core.content.edit
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.common.preferences.sharedPrefs
+import com.ichi2.anki.common.storage.CollectionHelper
+import com.ichi2.anki.common.storage.StorageDecision
 import com.ichi2.anki.exception.StorageNotConfiguredException
 import com.ichi2.anki.exception.SystemStorageException
-import com.ichi2.anki.storage.StorageDecision
 import org.junit.After
 import org.junit.Test
 import org.junit.runner.RunWith
