+++ anki-common/src/main/kotlin/com/ichi2/anki/exception/StorageNotConfiguredException.kt
@@ -2,9 +2,11 @@
 
 package com.ichi2.anki.exception
 
+import com.ichi2.anki.common.storage.StorageDecision
+
 /**
  * Thrown when the collection is accessed before the user has chosen where it should be stored
- * (the [storage decision][com.ichi2.anki.storage.StorageDecision] is `Undecided`).
+ * (the [storage decision][StorageDecision] is `Undecided`).
  *
  * Use `StorageAccessException` if a known path is unusable.
  */
